clarify NaN (numeric comparison) (#164)
not(!) part that was present before seemed to be a part of some unexplained syntax with ! sign (which is widely used in Julia), thus not being a clear description. While comparing two NaN's is not something that one would do, I believe comparing two matrices containing NaNs is a more widespread scenario and is important to know and be cautious about.